feat: implement QUNDIS HCA vendor extension with MbusValueDateG decoder
- Add complete QUNDIS HCA (Heat Cost Allocator) vendor extension
- Fix 10-year date offset bug in QUNDIS VIF 0x04 fields
- Add vendor registry with default extensions
- Enhance VIF parsing to route QUNDIS VIF 0x04 through vendor system
- Add comprehensive test suite with real device validation
- Add hardware acceleration support for Pi 5 deployment
- Add RTT logging for embedded debugging

+++ src/wmbus/radio/lora/single_channel.rs
@@ -184,9 +184,9 @@ impl SingleChannelConfig {
         let de = if ldro_enabled { 1 } else { 0 };
         let crc = if crc_on { 1 } else { 0 };
 
-        let payload_symb_nb = 8.0 + ((8 * payload_bytes as i32 - 4 * sf as i32
+        let payload_symb_nb = 8.0 + ((8 * payload_bytes as i32 - 4 * sf
             + 28 + 16 * crc - 20 * h) as f32
-            / (4 * (sf as i32 - 2 * de)) as f32).ceil() as f32
+            / (4 * (sf - 2 * de)) as f32).ceil()
             * (cr + 4) as f32;
 
         let payload_symb_nb = payload_symb_nb.max(0.0);
